requirements.txt
flask==2.0.1
gunicorn==20.1.0
langchain==0.1.0
langchain-community==0.0.10 
openai==1.3.0
PyPDF2==3.0.1
python-dotenv==0.19.0
faiss-cpu==1.7.4
tiktoken==0.5.2
werkzeug==2.0.1
